member DB Entity Test FIN

+++ src/main/java/com/nastech/upmureport/domain/repository/MemberRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nastech.upmureport.domain.entity.Member;
 
-public interface MemberRepository extends JpaRepository<Member, BigInteger> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 }
